requirements.txt
certifi==2025.1.31
influxdb-client==1.48.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
reactivex==4.0.4
setuptools==75.8.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
